improve: refactor directory components
