+++ app/lib/src/pages/badge/badge_controller.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import '../../components/main/badge/app_badge_base_builder.dart';
-import '../../components/main/topAppBar/app_top_bar_base_builder.dart';
+import '../../components/main/appBar/topAppBar/app_top_bar_base_builder.dart';
 import '../../config/app_theme_ext.dart';
 
 part 'badge_binding.dart';
